Test the behavior of sequence truncation
Add three sequences that are truncated at the tenth residue. The first
two are more than 10 residues long to test the correct truncation
length and the insertion of the <eos> token in the end. The last is
shorter than 10 residues to test the correct padding and position of
the <eos> token.