Fix: move sequence truncation into BatchConverter (#278)
* Fix: move sequence truncation into BatchConverter

The sequence truncation was performed in the extract.py. This led to
inconsistent sequences without <eos> tag for sequences longer than
1022. Moving the truncation into the BatchConverter prevents
inconsistency.

Resolves: #157

* Amend changes to test_readme

There was no reason to modify the Readme's test

* Amend changes to predict.py

* Change truncate bool into optional integer

Changing the truncate boolean into an optional integer with a default
value of None enables the user to specify the desired maximum length of
the sequence if any.

* Test the behavior of sequence truncation

Add three sequences that are truncated at the tenth residue. The first
two are more than 10 residues long to test the correct truncation
length and the insertion of the <eos> token in the end. The last is
shorter than 10 residues to test the correct padding and position of
the <eos> token.

* Amend changes to test_readme

There was no reason to modify the Readme's test

* Amend changes to variant-prediction/predict

* Amend changes to extract

* Add a parser argument to truncate the sequences

The addition of the truncation logic in the BatchConverter class
requires an additional parser argument to specify the maximum sequence
length. The `--truncation_seq_length` makes the old `--truncate`
argument obsolete and thus it is removed.

Co-authored-by: Alessandro Diana <69928507+AleD1996@users.noreply.github.com>